test(subscriptions): isolate group version hook
