+++ internal/cli/subscriptions/test_hooks.go
@@ -66,17 +66,3 @@ func SetSubscriptionVersionImageUploaderForTesting(fn func(context.Context, *os.
 		subscriptionVersionImageUploaderMu.Unlock()
 	}
 }
-
-// SetGroupVersionClientFactory replaces the group-version ASC client factory for tests.
-// It returns a restore function to reset the previous factory.
-func SetGroupVersionClientFactory(fn func() (*asc.Client, error)) func() {
-	previous := subscriptionGroupVersionClientFactory
-	if fn == nil {
-		subscriptionGroupVersionClientFactory = shared.GetASCClient
-	} else {
-		subscriptionGroupVersionClientFactory = fn
-	}
-	return func() {
-		subscriptionGroupVersionClientFactory = previous
-	}
-}
